+++ src/Control/Lens/Wrapped.hs
@@ -57,7 +57,6 @@ import           Control.Arrow
 import           Control.Applicative.Backwards
 import           Control.Comonad.Trans.Traced
 import           Control.Exception
-import           Control.Lens.Internal.Review
 import           Control.Lens.Iso
 import           Control.Lens.Review
 import           Control.Monad.Trans.Cont
@@ -159,12 +158,6 @@ instance Wrapped (Maybe a) (Maybe b) (Last a) (Last b) where
 
 instance (ArrowApply m, ArrowApply n) => Wrapped (m () a) (n () b) (ArrowMonad m a) (ArrowMonad n b) where
   wrapped = iso ArrowMonad getArrowMonad
-  {-# INLINE wrapped #-}
-
--- * lens
-
-instance Wrapped a b (Reviewed s a) (Reviewed t b) where
-  wrapped = iso Reviewed runReviewed
   {-# INLINE wrapped #-}
 
 -- * transformers