+++ CHANGELOG.markdown
@@ -8,6 +8,7 @@
 * Changed the signature of `Data.List.Split.Lens.condensing` due to the addition of `DropBlankFields` to `Data.List.Split.CondensePolicy` in `split`.
 * Replaced the use of `Accessor` with `Const`.
 * Replaced the use of `Mutator` with `Identity`.
+* Replaced the use of `Reviewed` with `Tagged`.
 
 3.9.0.2
 -------